Rename GearsBotNew example to GearsBot
