+++ wpilibcExamples/src/main/cpp/examples/GearsBot/cpp/commands/Pickup.cpp
