+++ wpilibcExamples/src/main/cpp/examples/GearsBot/include/commands/Pickup.h
