+++ wpilibcExamples/src/main/cpp/examples/GearsBot/include/commands/PrepareToPickup.h
