+++ wpilibjExamples/src/main/java/edu/wpi/first/wpilibj/examples/gearsbot/commands/Place.java
@@ -5,31 +5,30 @@
 /* the project.                                                               */
 /*----------------------------------------------------------------------------*/
 
-package edu.wpi.first.wpilibj.examples.gearsbotnew.commands;
+package edu.wpi.first.wpilibj.examples.gearsbot.commands;
 
 
 import edu.wpi.first.wpilibj2.command.SequentialCommandGroup;
 
-import edu.wpi.first.wpilibj.examples.gearsbotnew.subsystems.Claw;
-import edu.wpi.first.wpilibj.examples.gearsbotnew.subsystems.Elevator;
-import edu.wpi.first.wpilibj.examples.gearsbotnew.subsystems.Wrist;
+import edu.wpi.first.wpilibj.examples.gearsbot.subsystems.Claw;
+import edu.wpi.first.wpilibj.examples.gearsbot.subsystems.Elevator;
+import edu.wpi.first.wpilibj.examples.gearsbot.subsystems.Wrist;
 
 /**
- * Pickup a soda can (if one is between the open claws) and get it in a safe state to drive around.
+ * Place a held soda can onto the platform.
  */
-public class Pickup extends SequentialCommandGroup {
+public class Place extends SequentialCommandGroup {
   /**
-   * Create a new pickup command.
+   * Create a new place command.
    *
    * @param claw     The claw subsystem to use
    * @param wrist    The wrist subsystem to use
    * @param elevator The elevator subsystem to use
    */
-  public Pickup(Claw claw, Wrist wrist, Elevator elevator) {
+  public Place(Claw claw, Wrist wrist, Elevator elevator) {
     addCommands(
-        new CloseClaw(claw),
-        parallel(
-            new SetWristSetpoint(-45, wrist),
-            new SetElevatorSetpoint(0.25, elevator)));
+        new SetElevatorSetpoint(0.25, elevator),
+        new SetWristSetpoint(0, wrist),
+        new OpenClaw(claw));
   }
 }
