+++ wpilibjExamples/src/main/java/edu/wpi/first/wpilibj/examples/gearsbot/subsystems/Wrist.java
@@ -5,7 +5,7 @@
 /* the project.                                                               */
 /*----------------------------------------------------------------------------*/
 
-package edu.wpi.first.wpilibj.examples.gearsbotnew.subsystems;
+package edu.wpi.first.wpilibj.examples.gearsbot.subsystems;
 
 import edu.wpi.first.wpilibj.AnalogPotentiometer;
 import edu.wpi.first.wpilibj.Victor;
@@ -13,40 +13,37 @@ import edu.wpi.first.wpilibj.controller.PIDController;
 import edu.wpi.first.wpilibj.smartdashboard.SmartDashboard;
 import edu.wpi.first.wpilibj2.command.PIDSubsystem;
 
-import edu.wpi.first.wpilibj.examples.gearsbotnew.Robot;
+import edu.wpi.first.wpilibj.examples.gearsbot.Robot;
 
 /**
- * The elevator subsystem uses PID to go to a given height. Unfortunately, in it's current state PID
- * values for simulation are different than in the real world do to minor differences.
+ * The wrist subsystem is like the elevator, but with a rotational joint instead of a linear joint.
  */
-public class Elevator extends PIDSubsystem {
+public class Wrist extends PIDSubsystem {
   private final Victor m_motor;
   private final AnalogPotentiometer m_pot;
   private double m_setpoint;
 
-  private static final double kP_real = 4;
-  private static final double kI_real = 0.07;
-  private static final double kP_simulation = 18;
-  private static final double kI_simulation = 0.2;
+  private static final double kP_real = 1;
+  private static final double kP_simulation = 0.05;
 
   /**
-   * Create a new elevator subsystem.
+   * Create a new wrist subsystem.
    */
-  public Elevator() {
-    super(new PIDController(kP_real, kI_real, 0));
+  public Wrist() {
+    super(new PIDController(kP_real, 0, 0));
     if (Robot.isSimulation()) { // Check for simulation and update PID values
-      getController().setPID(kP_simulation, kI_simulation, 0);
+      getController().setPID(kP_simulation, 0, 0);
     }
-    getController().setAbsoluteTolerance(0.005);
+    getController().setAbsoluteTolerance(2.5);
 
-    m_motor = new Victor(5);
+    m_motor = new Victor(6);
 
     // Conversion value of potentiometer varies between the real world and
     // simulation
     if (Robot.isReal()) {
-      m_pot = new AnalogPotentiometer(2, -2.0 / 5);
+      m_pot = new AnalogPotentiometer(3, -270.0 / 5);
     } else {
-      m_pot = new AnalogPotentiometer(2); // Defaults to meters
+      m_pot = new AnalogPotentiometer(3); // Defaults to degrees
     }
 
     // Let's name everything on the LiveWindow
@@ -58,7 +55,7 @@ public class Elevator extends PIDSubsystem {
    * The log method puts interesting information to the SmartDashboard.
    */
   public void log() {
-    SmartDashboard.putData("Elevator Pot", m_pot);
+    SmartDashboard.putData("Wrist Angle", m_pot);
   }
 
   /**
